DROID-4201 Widgets | Refactor | Extract toSnackbarVariant to shared helper
Three identical copies of `toSnackbarVariant(space)` had been duplicated
across HomeScreenViewModel, ChatViewModel, and NewCreateObjectViewModel.
Promote it to a top-level suspend extension in UploadSuccessSnackbar.kt
that takes the StoreOfObjectTypes as a parameter, so future bug fixes
land in one place.

Behavior is unchanged.
